# docs/static/openapi.yaml
openapi: 3.0.0
info:
  contact:
    name: BearMount
    url: https://github.com/WhispersOfJ/bearmount/issues
  description: REST API for BearMount — Usenet WebDAV server with NZB queue,
    health monitoring, ARR integration, FUSE/rclone mounting, and Stremio addon.
  license:
    name: MIT
  termsOfService: https://github.com/WhispersOfJ/bearmount
  title: BearMount API
  version: "1.0"
paths:
  /arrs/download-client/register:
    post:
      description: Automatically registers BearMount as a download client
        (SABnzbd-compatible) in all configured ARR instances.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      security:
        - BearerAuth: []
      summary: Register ARR download clients
      tags:
        - ARRs
  /arrs/download-client/test:
    post:
      description: Tests whether BearMount is reachable as a download client from all
        configured ARR instances.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      security:
        - BearerAuth: []
      summary: Test ARR download clients
      tags:
        - ARRs
...
  /arrs/instances:
    get:
      description: Returns all configured ARR instances.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      security:
        - BearerAuth: []
      summary: List ARR instances
      tags:
        - ARRs
  "/arrs/instances/{type}/{name}":
    get:
      description: Returns a specific ARR instance by type and name.
      parameters:
        - description: Instance type (sonarr, radarr, lidarr, readarr, whisparr, or
            sportarr)
          in: path
          name: type
          required: true
          schema:
            type: string
        - description: Instance name
          in: path
          name: name
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      security:
        - BearerAuth: []
      summary: Get ARR instance
      tags:
        - ARRs
  /arrs/instances/test:
    post:
      description: Tests connectivity to an ARR instance with given credentials.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/api.TestConnectionRequest"
        description: Instance connection details
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      security:
        - BearerAuth: []
      summary: Test ARR connection
      tags:
        - ARRs
  /arrs/stats:
    get:
      description: Returns sync statistics for all configured ARR instances.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      security:
        - BearerAuth: []
      summary: Get ARR statistics
      tags:
        - ARRs
  /arrs/webhook:
    post:
      description: Receives file-import webhook events from ARR instances and triggers
        health checks. Authenticated via API key query parameter.
      parameters:
        - description: BearMount API key
          in: query
          name: apikey
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/api.ArrsWebhookRequest"
        description: Webhook payload
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      security:
        - ApiKeyAuth: []
      summary: ARR webhook receiver
      tags:
        - ARRs
  /arrs/webhook/register:
    post:
      description: Automatically registers BearMount as a webhook connection in all
        configured ARR instances.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      security:
        - BearerAuth: []
      summary: Register ARR webhooks
      tags:
        - ARRs
...
  /auth/login:
    post:
      description: Authenticates with username and password and returns a JWT access
        token. Rate-limited to 10/min per IP.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/api.LoginRequest"
        description: Login credentials
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/api.APIResponse"
                  - properties:
                      data:
                        $ref: "#/components/schemas/api.AuthResponse"
                    type: object
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
        "429":
          description: Too Many Requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      summary: Login
      tags:
        - Auth
  /auth/register:
    post:
      description: Creates the first admin user account. Only allowed when no users
        exist yet.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/api.RegisterRequest"
        description: Registration details
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/api.APIResponse"
                  - properties:
                      data:
                        $ref: "#/components/schemas/api.AuthResponse"
                    type: object
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      summary: Register
      tags:
        - Auth
...
  /auth/reset-admin-password:
    post:
      description: Resets the password for a direct-auth user. Only available when
        login is disabled.
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/api.ResetAdminPasswordRequest"
        description: Reset credentials
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      summary: Reset admin password
      tags:
        - Auth
...
  "/config/{section}":
    patch:
      description: Updates a specific named section of the configuration (e.g. import,
        sabnzbd, rclone).
      parameters:
        - description: Config section name
          in: path
          name: section
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
        description: Section configuration
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      security:
        - BearerAuth: []
      summary: Patch configuration section
      tags:
        - Config
...
  /config/reload:
    post:
      description: Reloads the BearMount configuration from disk without restarting.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api.APIResponse"
      security:
        - BearerAuth: []
      summary: Reload configuration
      tags:
        - Config
...
servers:
  - url: //localhost:8080/api
components:
...
  securitySchemes:
    ApiKeyAuth:
      in: query
      name: apikey
      type: apiKey
  schemas:
    api.APIError:
      properties:
        code:
          type: string
        details:
          type: string
        message:
          type: string
      type: object
...
    api.APIResponse:
      properties:
        data: {}
        error:
          $ref: "#/components/schemas/api.APIError"
        meta:
          $ref: "#/components/schemas/api.APIMeta"
        success:
          type: boolean
      type: object
...
    api.AuthResponse:
      properties:
        message:
          type: string
        redirect_url:
          type: string
        user:
          $ref: "#/components/schemas/api.UserResponse"
      type: object
...
    api.LoginRequest:
      properties:
        password:
          type: string
        username:
          type: string
      type: object
...
    api.RegisterRequest:
      properties:
        email:
          type: string
        password:
          type: string
        username:
          type: string
      type: object
    api.ResetAdminPasswordRequest:
      properties:
        new_password:
          type: string
        username:
          type: string
      type: object
...
    api.TestConnectionRequest:
      properties:
        api_key:
          type: string
        type:
          type: string
        url:
          type: string
      type: object
...